requirements.txt
django
djangorestframework
markdown
django-filter
django-rest-knox
django-cors-headers